Add different styles based on energy amount
When the creature has enough energy available, the part of the bar
being consumed of the total available has a yellow and black
horizontal stripe pattern. Any remaining energy is shown as proper
yellow.

When the creature doesn't have enough energy available, the current
amount is at the bottom of the bar with 25% opacity yellow. On top of
that is a black bar with a yellow outline indicating how much more
energy is required.