DEBUG .env variables
Bugfix: values in .env are substitued into source as quoted strings.
But in game.js, debugMode treated it as a boolean. `Boolean("false")` in
JS is `true`, so debugMode was always on.

This commit also adds two new .env key/values:

* `DEBUG_DISABLE_GAME_STATUS_CONSOLE_LOG=true` to disable default console
logging of game status.
* `DEBUG_DISABLE_MUSIC=true` to turn off music.

In both cases, if DEBUG_MODE is set to anything but "true" the values
are ignored.