Bounty hunter ability and pierce targeting
Added new options for when checking for a target creature and for when targeting.

pierceThroughBehavior: Set to "stop", "partial", or "pierce".
Stop means that there is no piercing effect, and if the creature does not have a direct line of sight, an enemy is out of range.
Pierce goes through everything, and only stops after it's seen peirceNumber valid targets.
Partial will pierce targets till it encounters an invalid target, and then stop. (When looking for an enemy, partial pierces creatures of the target team who fail the optional tests to be considered a target)

Fixed upgraded shotgun shot being usable the same turn as the upgrade.